model testing done

+++ migrations/sqls/20230305220003-orders-table-up.sql
@@ -1,8 +1,8 @@
 /* Replace with your SQL commands */
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
-    product_id bigint REFERENCES products(id) NOT NULL,
+    product_id INTEGER REFERENCES products(id) NOT NULL,
     product_qty INTEGER NOT NULL,
-    user_id bigint REFERENCES users(id),
+    user_id INTEGER REFERENCES users(id),
     order_status VARCHAR NOT NULL    
 );
